Refactor using-contentful to use fs routes
